+++ src/get-angle/index.js
@@ -1,5 +1,4 @@
 import { getCoord } from '@turf/invariant'
-import { round } from '@turf/helpers'
 
 /**
  * Finds the angle between 3 points.
@@ -7,8 +6,6 @@ import { round } from '@turf/helpers'
  * @param {number[]} startPoint Start Point Coordiantes
  * @param {number[]} midPoint Mid Point Coordinates
  * @param {number[]} endPoint End Point Coordinates
- * @param {Object} [options] Optional parameters
- * @param {number} [options.precision] Number of precison
  * @return {number} angle
  * @example
  * osmlinter.getAngle([5, 5], [5, 6], [3, 4])
@@ -16,7 +13,6 @@ import { round } from '@turf/helpers'
  */
 export default function (startPoint, midPoint, endPoint, options) {
   options = options || {}
-  var precision = options.precision || 6
 
   // Rename to shorter variables
   var A = getCoord(startPoint)
@@ -29,5 +25,5 @@ export default function (startPoint, midPoint, endPoint, options) {
   var BC = Math.sqrt(Math.pow(B[0] - C[0], 2) + Math.pow(B[1] - C[1], 2))
   var AC = Math.sqrt(Math.pow(C[0] - A[0], 2) + Math.pow(C[1] - A[1], 2))
   var angle = Math.acos((BC * BC + AB * AB - AC * AC) / (2 * BC * AB)) * (180 / pi)
-  return round(angle, precision)
+  return angle
 }